requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py36h06a4308_0
chardet=4.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
einops=0.4.0=pypi_0
expat=2.3.0=h2531618_2
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
glib=2.68.1=h36276a3_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
google-auth=1.28.0=pypi_0
google-auth-oauthlib=0.4.3=pypi_0
grpcio=1.36.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pypi_0
importlib-metadata=3.7.3=pypi_0
intel-openmp=2020.2=254
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h3942068_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
opencv-python=4.5.1.48=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
pcre=8.44=he6710b0_0
pillow=8.1.2=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
protobuf=3.15.6=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.8.0=py3.6_cuda10.2_cudnn7.6.5_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.5.4=pypi_0
setuptools=52.0.0=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py36h06a4308_0
sqlite=3.35.1=hdfb4753_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.1=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
torch-dct=0.1.5=pypi_0
torchvision=0.9.0=py36_cu102
tornado=6.1=py36h27cfd23_0
tqdm=4.59.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
